Began shortening tall glyphs.
Started effort to reduce tall symbols and capitals by 100px in height.

Doing the following is a quick way to determine which glyphs need to be changed:

    s/"y" : -600/"y" : -501/g
    s/"y" : -350/"y" : -251/g

Then build loose.

If a glyph needs changed, replace -501 with -500 or -251 with -200 before saving. If not, leave it, then reverse the substitutions when finished editing.

The new "1" contains a line that is only used for spacing purposes and contains `"draw" : false`. At some point the JSON-to-loose converter will recognize this and include the line for measuring the width of the glyph, but omit it from the drawn paths. This would also be a good way to implement the space glyph, which is currently hardcoded into the FontForge export. The export process will need to make sure there is at least one drawable path for a glyph. If not, it should skip the SVG import (which I've read fails on an empty file).