Support for spacing-only lines.
A line can now be specified as `"draw" : false` (and the `guide_color` property is made optional). Setting this to false allows the line to be counted when calculating the bounding box (including, importantly, the width) but causes it to be omitted when drawing strokes in the SVG.

`elements-all.json` now contains a glyph for space.

The font project generator skips importing a spacing-only SVG since it would be a no-op anyway (and seems to throw up an error message due to the lack of paths).

The master reformatter puts `draw` just before `guide_color` in the sort order.