Add `spread` property.
Essentially a counter to the `draw` property, a line or compose item may be marked `"spread" : false`. If false, the line (for a line) or result lines (for a compose item) do not affect the containing glyph's bounding box (i.e., width) but are still drawn (unless `"draw" : false`, which is probably worthless).