Composition transforms now in `ComposeRunner.pm`.
The original `get_compose_lines` in `element-to-loose-svg.pl` was growing into a hairy dispatch table. It is now implemented more naturally as a state object. Methods can be looked up and called by a string name, so a JSON array can easily
be converted to a call with parameter list.