Composer taught to use anchors.
A short sequence of compose commands can be used with anchors to properly place and center diacritics.

In this commit is an example for `idieresis`. An anchor `d-top` is specified for the dieresis (it is placed 50 below the bottom-center of the diacritic). Compose instructions are used to align the `d-top` from the dieresis with the `top` from the letter. (`top` and a few others are automatically set to bounding box edges unless specified explicitly.)