Added `aligndia` compose op.
`["aligndia","NAME"]` is a shorthand for translating the current glyph so that the target `d-NAME` is at the same coordinates as the target `NAME` from a previously drawn compose item `base`. For example, in `circumflex` we define `d-top` at 50 below the bottom center of that glyph. When composing `ecircumflex`, we include `e` and name the compose item `base`, then we include `circumflex` and use the op `["aligndia","top"]`. `d-top` of this glyph is now aligned with `top` of the base glyph.